Sentiment Analysis of IU-related Twitter mentions
Sentiment Analysis of Twitter accounts to uncover sentiment toward IU
Sentiment analysis comes in three major categrories: a positive,
negative or neutral sentiment on the specific topic.

This Python program is created to determine the sentiment of a writer
from a twitter account within an online mention of tweets on topis
related to IU (Indiana University). Four specific keywords will be used
to get the sentiment on them based on the latest 2000 tweets. The code
to collect the corpus of tweets was created based on Al-Masri (2019),
Chauhan (2018)